Move ParticlePhysics in it's own folder
