+++ Core/include/Voxymore/ParticlesPhysics/Components/ParticleComponent.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "Voxymore/Components/CustomComponent.hpp"
-#include "Voxymore/Physics/Particles/Particle.hpp"
+#include "Voxymore/ParticlePhysic/Particle.hpp"
 #include "Voxymore/Scene/Entity.hpp"
 
 namespace Voxymore::Core
