+++ Core/include/Voxymore/ParticlesPhysics/ParticlePhysicsLayer.hpp
@@ -16,8 +16,8 @@
 #include "Voxymore/Core/FileSystem.hpp"
 #include "Voxymore/Scene/Scene.hpp"
 
-#include "Voxymore/Physics/Particles/Collisions/ParticleContact.hpp"
-#include "Voxymore/Physics/Particles/Collisions/ParticleContactResolver.hpp"
+#include "Voxymore/ParticlePhysic/Collisions/ParticleContact.hpp"
+#include "Voxymore/ParticlePhysic/Collisions/ParticleContactResolver.hpp"
 
 #include "Voxymore/Physics/Rigidbodies/Components/RigidbodyComponent.hpp"
 #include "Voxymore/Physics/Rigidbodies/Systems/RBGravitySystem.hpp"
