+++ Core/include/Voxymore/Voxymore.hpp
@@ -36,12 +36,12 @@
 #include "Voxymore/Components/ModelComponent.hpp"
 #include "Voxymore/Components/PrimitiveComponent.hpp"
 
+#include "Voxymore/ParticlePhysic/Particle.hpp"
+#include "Voxymore/ParticlePhysic/ParticlePhysicsLayer.hpp"
 #include "Voxymore/Physics/Particles/Collisions/ParticleContact.hpp"
 #include "Voxymore/Physics/Particles/Components/AnchoredSpringComponent.hpp"
 #include "Voxymore/Physics/Particles/Components/FloatingComponent.hpp"
 #include "Voxymore/Physics/Particles/Components/ParticleComponent.hpp"
-#include "Voxymore/Physics/Particles/Particle.hpp"
-#include "Voxymore/Physics/Particles/ParticlePhysicsLayer.hpp"
 #include "Voxymore/Physics/Particles/Systems/BuoyancyForceSystem.hpp"
 #include "Voxymore/Physics/Particles/Systems/SpringForceSystem.hpp"
 #include "Voxymore/Physics/Rigidbodies/Components/RigidbodyComponent.hpp"
